Big Multiplayer Refactoring
